refactor(server): split bootstrap into server.ts, dedupe room lookup
Follows the harness package's convention of a pure index.ts export
surface with process/bootstrap concerns in their own entrypoint
(dist/server.js is what the Pi hosting research doc's systemd unit
expects). Also extracts the repeated "look up room, 404 if missing"
block from the join/qr/end routes into one helper.
